refactor: remove unnecessary abstractions and tests Many tests were just testing the hand crafted stubs.
