+++ provider/gce/google/config_credentials_test.go
@@ -12,10 +12,11 @@ import (
 	gc "gopkg.in/check.v1"
 
 	"github.com/juju/juju/provider/gce/google"
+	"github.com/juju/juju/testing"
 )
 
 type credentialsSuite struct {
-	google.BaseSuite
+	testing.BaseSuite
 }
 
 var _ = gc.Suite(&credentialsSuite{})
