+++ provider/gce/google/conn.go
@@ -5,123 +5,20 @@ package google
 
 import (
 	"context"
+	"fmt"
+	"net/http"
+	"path"
+	"time"
 
+	"github.com/juju/clock"
 	"github.com/juju/errors"
 	jujuhttp "github.com/juju/http/v2"
+	"github.com/juju/retry"
 	"google.golang.org/api/compute/v1"
+	"google.golang.org/api/googleapi"
+	"google.golang.org/api/option"
+	transporthttp "google.golang.org/api/transport/http"
 )
-
-// service facilitates mocking out the GCE API during tests.
-// TODO (manadart 2019-11-19): This indirection is for *our* wrapper of the
-// underlying compute services. As an indirection, it is one layer too high
-// and sits on an unnecessary type abstraction.
-// We should:
-//   - Create interfaces for members of the compute service so our line of
-//     indirection is at the boundary with the dependency.
-//   - Use the compute service as the implementer of those interfaces and
-//     embed them directly in an equivalent of Connection (below).
-//   - Remove this interface and rawConn altogether.
-//   - Remove types that have no purpose other than to mirror those returned by
-//     the compute service such as `MachineType`.
-type service interface {
-	// GetProjectServiceAccount sends a request to the GCE API to get
-	// the service account for the specified project.
-	// If the project does not exist then an error will be returned.
-	GetProjectServiceAccount(projectID string) (string, error)
-
-	// GetInstance sends a request to the GCE API for info about the
-	// specified instance. If the instance does not exist then an error
-	// will be returned.
-	GetInstance(projectID, id, zone string) (*compute.Instance, error)
-
-	// ListInstances sends a request to the GCE API for a list of all
-	// instances in project for which the name starts with the provided
-	// prefix. The result is also limited to those instances with one of
-	// the specified statuses (if any).
-	ListInstances(projectID, prefix string, status ...string) ([]*compute.Instance, error)
-
-	// AddInstance sends a request to GCE to add a new instance to the
-	// given project, with the provided instance data. The call blocks
-	// until the instance is created or the request fails.
-	AddInstance(projectID, zone string, spec *compute.Instance) error
-
-	// RemoveInstance sends a request to the GCE API to remove the instance
-	// with the provided ID (in the specified zone). The call blocks until
-	// the instance is removed (or the request fails).
-	RemoveInstance(projectID, id, zone string) error
-
-	// SetMetadata sends a request to the GCE API to update one
-	// instance's metadata. The call blocks until the request is
-	// completed or fails.
-	SetMetadata(projectID, zone, instanceID string, metadata *compute.Metadata) error
-
-	// GetFirewalls sends an API request to GCE for the information about
-	// the firewalls with the namePrefix and returns them.
-	// If no firewalls are not found, errors.NotFound is returned.
-	GetFirewalls(projectID, namePrefix string) ([]*compute.Firewall, error)
-
-	// AddFirewall requests GCE to add a firewall with the provided info.
-	// If the firewall already exists then an error will be returned.
-	// The call blocks until the firewall is added or the request fails.
-	AddFirewall(projectID string, firewall *compute.Firewall) error
-
-	// UpdateFirewall requests GCE to update the named firewall with the
-	// provided info, overwriting the existing data. If the firewall does
-	// not exist then an error will be returned. The call blocks until the
-	// firewall is updated or the request fails.
-	UpdateFirewall(projectID, name string, firewall *compute.Firewall) error
-
-	// RemoveFirewall removes the named firewall from the project. If it
-	// does not exist then this is a noop. The call blocks until the
-	// firewall is added or the request fails.
-	RemoveFirewall(projectID, name string) error
-
-	// ListAvailabilityZones returns the list of availability zones for a given
-	// GCE region. If none are found the the list is empty. Any failure in
-	// the low-level request is returned as an error.
-	ListAvailabilityZones(projectID, region string) ([]*compute.Zone, error)
-
-	// CreateDisk will create a gce Persistent Block device that matches
-	// the specified in spec.
-	CreateDisk(project, zone string, spec *compute.Disk) error
-
-	// ListDisks returns a list of disks available for a given project.
-	ListDisks(project string) ([]*compute.Disk, error)
-
-	// RemoveDisk will delete the disk identified by id.
-	RemoveDisk(project, zone, id string) error
-
-	// GetDisk will return the disk correspondent to the passed id.
-	GetDisk(project, zone, id string) (*compute.Disk, error)
-
-	// SetDiskLabels sets the labels on a disk, ensuring that the disk's
-	// label fingerprint matches the one supplied.
-	SetDiskLabels(project, zone, id, labelFingerprint string, labels map[string]string) error
-
-	// AttachDisk will attach the disk described in attachedDisks (if it exists) into
-	// the instance with id instanceId.
-	AttachDisk(project, zone, instanceId string, attachedDisk *compute.AttachedDisk) error
-
-	// Detach disk detaches device diskDeviceName (if it exists and its attached)
-	// form the machine with id instanceId.
-	DetachDisk(project, zone, instanceId, diskDeviceName string) error
-
-	// InstanceDisks returns the disks attached to the instance identified
-	// by instanceId
-	InstanceDisks(project, zone, instanceId string) ([]*compute.AttachedDisk, error)
-
-	// ListMachineTypes returns a list of machines available in the project and zone provided.
-	ListMachineTypes(projectID, zone string) (*compute.MachineTypeList, error)
-
-	// ListSubnetworks returns a list of subnets available in the given project and region.
-	ListSubnetworks(projectID, region string) ([]*compute.Subnetwork, error)
-
-	// ListNetworks returns a list of Networks available in the given project.
-	ListNetworks(projectID string) ([]*compute.Network, error)
-}
-
-// TODO(ericsnow) Add specific error types for common failures
-// (e.g. BadRequest, RequestFailed, RequestError, ConnectionFailed)?
 
 // Connection provides methods for interacting with the GCE API. The
 // methods are limited to those needed by the juju GCE provider.
@@ -130,7 +27,7 @@ type service interface {
 // called to authenticate and open the raw connection to the GCE API.
 // Otherwise a panic will result.
 type Connection struct {
-	service   service
+	*compute.Service
 	projectID string
 }
 
@@ -140,45 +37,208 @@ type Connection struct {
 // result in an error. All errors that happen while authenticating and
 // connecting are returned by Connect.
 func Connect(ctx context.Context, connCfg ConnectionConfig, creds *Credentials) (*Connection, error) {
-	raw, err := newService(ctx, creds, connCfg.HTTPClient)
+	service, err := newComputeService(ctx, creds, connCfg.HTTPClient)
 	if err != nil {
 		return nil, errors.Trace(err)
 	}
 
 	conn := &Connection{
-		service:   &rawConn{raw},
+		Service:   service,
 		projectID: connCfg.ProjectID,
 	}
 	return conn, nil
 }
 
-var newService = func(ctx context.Context, creds *Credentials, httpClient *jujuhttp.Client) (*compute.Service, error) {
-	return newComputeService(ctx, creds, httpClient)
+// newComputeService opens a new low-level connection to the GCE API using
+// the input credentials and returns it.
+// This includes building the OAuth-wrapping network transport.
+func newComputeService(ctx context.Context, creds *Credentials, httpClient *jujuhttp.Client) (*compute.Service, error) {
+	cfg, err := newJWTConfig(creds)
+	if err != nil {
+		return nil, errors.Trace(err)
+	}
+
+	// We're substituting the transport, with a wrapped GCE specific version of
+	// the original http.Client.
+	newClient := *httpClient.Client()
+
+	tsOpt := option.WithTokenSource(cfg.TokenSource(ctx))
+	if newClient.Transport, err = transporthttp.NewTransport(ctx, newClient.Transport, tsOpt); err != nil {
+		return nil, errors.Trace(err)
+	}
+
+	service, err := compute.NewService(ctx,
+		tsOpt,
+		option.WithHTTPClient(&newClient),
+	)
+	return service, errors.Trace(err)
+}
+
+func (c *Connection) getProjectServiceAccount(ctx context.Context, projectID string) (string, error) {
+	call := c.Projects.Get(projectID).Fields("defaultServiceAccount").
+		Context(ctx)
+	proj, err := call.Do()
+	if err != nil {
+		return "", errors.Trace(err)
+	}
+	return proj.DefaultServiceAccount, nil
 }
 
 // VerifyCredentials ensures that the authentication credentials used
 // to connect are valid for use in the project and region defined for
 // the Connection. If they are not then an error is returned.
-func (gc Connection) VerifyCredentials() error {
-	if _, err := gc.service.GetProjectServiceAccount(gc.projectID); err != nil {
+func (c *Connection) VerifyCredentials(ctx context.Context) error {
+	if _, err := c.getProjectServiceAccount(ctx, c.projectID); err != nil {
 		// TODO(ericsnow) Wrap err with something about bad credentials?
 		return errors.Trace(err)
 	}
 	return nil
 }
 
-// AvailabilityZones returns the list of availability zones for a given
-// GCE region. If none are found the the list is empty. Any failure in
-// the low-level request is returned as an error.
-func (gc *Connection) AvailabilityZones(region string) ([]AvailabilityZone, error) {
-	rawZones, err := gc.service.ListAvailabilityZones(gc.projectID, region)
-	if err != nil {
-		return nil, errors.Trace(err)
+// DefaultServiceAccount returns the default service account for the project.
+func (c *Connection) DefaultServiceAccount(ctx context.Context) (string, error) {
+	return c.getProjectServiceAccount(ctx, c.projectID)
+}
+
+// These are attempt strategies used in waitOperation.
+var (
+	// TODO(ericsnow) Tune the timeouts and delays.
+	longRetryStrategy = retry.CallArgs{
+		Clock:       clock.WallClock,
+		Delay:       2 * time.Second,
+		MaxDuration: 5 * time.Minute,
+	}
+)
+
+func convertRawAPIError(err error) error {
+	if err2, ok := err.(*googleapi.Error); ok {
+		if err2.Code == http.StatusNotFound {
+			return errors.NewNotFound(err, "")
+		}
+	}
+	return err
+}
+
+type handleOperationErrors func(operation *compute.Operation) error
+
+func returnNotFoundOperationErrors(operation *compute.Operation) error {
+	if operation.Error != nil {
+		result := waitError{operation, nil}
+		for _, err := range operation.Error.Errors {
+			if err.Code == "RESOURCE_NOT_FOUND" {
+				result.cause = errors.NotFoundf("%v: resource", err.Message)
+				continue
+			}
+			logger.Errorf("GCE operation error: (%s) %s", err.Code, err.Message)
+		}
+		return result
+	}
+	return nil
+}
+
+func logOperationErrors(operation *compute.Operation) error {
+	if operation.Error != nil {
+		for _, err := range operation.Error.Errors {
+			logger.Errorf("GCE operation error: (%s) %s", err.Code, err.Message)
+		}
+		return waitError{operation, nil}
+	}
+	return nil
+}
+
+type waitError struct {
+	op    *compute.Operation
+	cause error
+}
+
+func (err waitError) Error() string {
+	if err.cause != nil {
+		return fmt.Sprintf("GCE operation %q failed: %v", err.op.Name, err.cause)
+	}
+	return fmt.Sprintf("GCE operation %q failed", err.op.Name)
+}
+
+func (err waitError) Cause() error {
+	if err.cause != nil {
+		return err.cause
+	}
+	return err
+}
+
+func isWaitError(err error) bool {
+	_, ok := err.(*waitError)
+	return ok
+}
+
+type opDoer interface {
+	Do(...googleapi.CallOption) (*compute.Operation, error)
+}
+
+// checkOperation requests a new copy of the given operation from the
+// GCE API and returns it. The new copy will have the operation's
+// current status.
+func (c *Connection) checkOperation(projectID string, op *compute.Operation) (*compute.Operation, error) {
+	var call opDoer
+	if op.Zone != "" {
+		zoneName := path.Base(op.Zone)
+		call = c.ZoneOperations.Get(projectID, zoneName, op.Name)
+	} else if op.Region != "" {
+		region := path.Base(op.Region)
+		call = c.RegionOperations.Get(projectID, region, op.Name)
+	} else {
+		call = c.GlobalOperations.Get(projectID, op.Name)
 	}
 
-	var zones []AvailabilityZone
-	for _, rawZone := range rawZones {
-		zones = append(zones, AvailabilityZone{rawZone})
+	operation, err := doOpCall(call)
+	if err != nil {
+		return nil, errors.Annotatef(err, "request for GCE operation %q failed", op.Name)
 	}
-	return zones, nil
+	return operation, nil
+}
+
+var doOpCall = func(call opDoer) (*compute.Operation, error) {
+	return call.Do()
+}
+
+// waitOperation waits for the provided operation to reach the "done"
+// status. It follows the given attempt strategy (e.g. wait time between
+// attempts) and may time out.
+func (c *Connection) waitOperation(projectID string, op *compute.Operation, retryStrategy retry.CallArgs, f handleOperationErrors) error {
+	// TODO(perrito666) 2016-05-02 lp:1558657
+	started := time.Now()
+	logger.Infof("GCE operation %q, waiting...", op.Name)
+
+	retryStrategy.IsFatalError = func(err error) bool {
+		return !errors.IsNotProvisioned(err)
+	}
+	retryStrategy.Func = func() error {
+		var err error
+		op, err = c.checkOperation(projectID, op)
+		if err != nil {
+			return err
+		}
+		if op.Status == StatusDone {
+			return nil
+		}
+		return errors.NewNotProvisioned(nil, "GCE operation not done yet")
+	}
+	var err error
+	if op.Status != StatusDone {
+		err = retry.Call(retryStrategy)
+	}
+
+	if retry.IsAttemptsExceeded(err) || retry.IsDurationExceeded(err) {
+		// lp:1558657
+		err := errors.Annotatef(err, "timed out after %d seconds", time.Now().Sub(started)/time.Second)
+		return waitError{op, err}
+	}
+	if err != nil {
+		return errors.Trace(err)
+	}
+	if err := f(op); err != nil {
+		return err
+	}
+
+	logger.Infof("GCE operation %q finished", op.Name)
+	return nil
 }
