+++ provider/gce/google/network_test.go
@@ -1,113 +1,27 @@
-// Copyright 2014 Canonical Ltd.
+// Copyright 2025 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package google_test
+package google
 
 import (
-	"sort"
-
-	jc "github.com/juju/testing/checkers"
-	"google.golang.org/api/compute/v1"
 	gc "gopkg.in/check.v1"
-
-	"github.com/juju/juju/core/network"
-	"github.com/juju/juju/provider/gce/google"
 )
 
-type networkSuite struct {
-	google.BaseSuite
+type firewallNameSuite struct{}
+
+var _ = gc.Suite(&firewallNameSuite{})
+
+func (s *firewallNameSuite) TestSimplePattern(c *gc.C) {
+	res := matchesPrefix("juju-3-123", "juju-3")
+	c.Assert(res, gc.Equals, true)
 }
 
-var _ = gc.Suite(&networkSuite{})
-
-func (s *networkSuite) TestNetworkSpecPath(c *gc.C) {
-	spec := google.NetworkSpec{
-		Name: "spam",
-	}
-	path := spec.Path()
-
-	c.Check(path, gc.Equals, "global/networks/spam")
+func (s *firewallNameSuite) TestExactMatch(c *gc.C) {
+	res := matchesPrefix("juju-3", "juju-3")
+	c.Assert(res, gc.Equals, true)
 }
 
-func (s *networkSuite) TestNetworkSpecNewInterface(c *gc.C) {
-	spec := google.NetworkSpec{
-		Name: "spam",
-	}
-	netIF := google.NewNetInterface(spec, "eggs", true)
-
-	c.Check(netIF, gc.DeepEquals, &compute.NetworkInterface{
-		Network: "global/networks/spam",
-		AccessConfigs: []*compute.AccessConfig{{
-			Name: "eggs",
-			Type: google.NetworkAccessOneToOneNAT,
-		}},
-	})
-}
-
-type ByIPProtocol []*compute.FirewallAllowed
-
-func (s ByIPProtocol) Len() int {
-	return len(s)
-}
-func (s ByIPProtocol) Swap(i, j int) {
-	s[i], s[j] = s[j], s[i]
-}
-func (s ByIPProtocol) Less(i, j int) bool {
-	return s[i].IPProtocol < s[j].IPProtocol
-}
-
-func (s *networkSuite) TestFirewallSpec(c *gc.C) {
-	ports := map[string][]network.PortRange{
-		"tcp":  {{FromPort: 80, ToPort: 81}, {FromPort: 8888, ToPort: 8888}},
-		"udp":  {{FromPort: 1234, ToPort: 1234}},
-		"icmp": {{FromPort: -1, ToPort: -1}},
-	}
-	fw := google.FirewallSpec("spam", "target", []string{"192.168.1.0/24", "10.0.0.0/24"}, ports)
-
-	allowed := []*compute.FirewallAllowed{{
-		IPProtocol: "icmp",
-		Ports:      []string{},
-	}, {
-		IPProtocol: "tcp",
-		Ports:      []string{"80-81", "8888"},
-	}, {
-		IPProtocol: "udp",
-		Ports:      []string{"1234"},
-	}}
-	sort.Sort(ByIPProtocol(fw.Allowed))
-	for i := range fw.Allowed {
-		sort.Strings(fw.Allowed[i].Ports)
-	}
-	c.Check(fw, jc.DeepEquals, &compute.Firewall{
-		Name:         "spam",
-		TargetTags:   []string{"target"},
-		SourceRanges: []string{"192.168.1.0/24", "10.0.0.0/24"},
-		Allowed:      allowed,
-	})
-}
-
-func (s *networkSuite) TestExtractAddresses(c *gc.C) {
-	addresses := google.ExtractAddresses(&s.NetworkInterface)
-
-	c.Check(addresses, jc.DeepEquals, []network.ProviderAddress{
-		network.NewMachineAddress("10.0.0.1", network.WithScope(network.ScopeCloudLocal)).AsProviderAddress(),
-	})
-}
-
-func (s *networkSuite) TestExtractAddressesExternal(c *gc.C) {
-	s.NetworkInterface.NetworkIP = ""
-	s.NetworkInterface.AccessConfigs[0].NatIP = "8.8.8.8"
-	addresses := google.ExtractAddresses(&s.NetworkInterface)
-
-	c.Check(addresses, jc.DeepEquals, []network.ProviderAddress{
-		network.NewMachineAddress("8.8.8.8", network.WithScope(network.ScopePublic)).AsProviderAddress(),
-	})
-}
-
-func (s *networkSuite) TestExtractAddressesEmpty(c *gc.C) {
-	s.NetworkInterface.AccessConfigs = nil
-	s.NetworkInterface.NetworkIP = ""
-	addresses := google.ExtractAddresses(&s.NetworkInterface)
-
-	c.Check(addresses, gc.HasLen, 0)
+func (s *firewallNameSuite) TestThatJujuMachineIDsDoNotCollide(c *gc.C) {
+	res := matchesPrefix("juju-30-123", "juju-3")
+	c.Assert(res, gc.Equals, false)
 }
